requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
coverage==5.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cleanup==5.0.0
django-dbbackup==3.3.0
django-debug-toolbar==2.2
django-filebrowser==3.13.1
django-grappelli==2.14.2
django-import-export==2.2.0
django-timezone-field==4.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
flake8==3.8.3
future==0.18.2
jdcal==1.4.1
kombu==4.6.11
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
